@@ -65,7 +65,7 @@
     "nodemon": "^3.1.9",
     "ts-node": "^10.9.2",
     "tsc-alias": "^1.8.15",
-    "typescript": "^5.7.3"
+    "typescript": "^5.8.3"
   },
   "_moduleAliases": {
     "@": "./"
